requirements.txt
astroid==2.1.0
certifi==2018.11.29
chardet==3.0.4
icalendar==4.0.3
idna==2.8
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
numpy==1.16.1
pandas==0.24.1
ply==3.11
pprint==0.1
pylint==2.2.2
pypandoc==1.4
python-dateutil==2.8.0
python-gitlab==1.7.0
pytz==2018.9
requests==2.21.0
six==1.12.0
typed-ast==1.3.1
urllib3==1.24.1
wrapt==1.11.1
